+++ src/App.js
@@ -12,6 +12,7 @@ import { destroyStore } from './mainActions';
 import css from './components/SessionEventContainer/style.css';
 
 import Root from './components/Root';
+import { eventsPortal } from './constants';
 
 const StrictWrapper = ({ children }) => {
   if (config.disableStrictMode) {
@@ -64,7 +65,7 @@ export default class StripesCore extends Component {
     return (
       <StrictWrapper>
         <div
-          id="events-portal"
+          id={eventsPortal}
           className={css.eventsContainer}
         />
         <Root