+++ src/RootWithIntl.js
@@ -1,6 +1,5 @@
 import { useState } from 'react';
 import PropTypes from 'prop-types';
-import { createPortal } from 'react-dom';
 import {
   Router,
   Switch,
@@ -76,11 +75,7 @@ const RootWithIntl = ({ stripes, token = '', isAuthenticated = false, disableAut
                           {(typeof connectedStripes.okapi !== 'object' || connectedStripes.discovery.isFinished) && (
                             <ModuleContainer id="content">
                               <OverlayContainer />
-                              {connectedStripes.config.useSecureTokens &&
-                                createPortal(<SessionEventContainer
-                                  history={history}
-                                />,
-                                document.getElementById('events-portal'))}
+                              <SessionEventContainer history={history} />
                               <Switch>
                                 <TitledRoute
                                   name="home"